% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oml_build.R
\name{oml_build}
\alias{oml_build}
\title{R wrapper for ./gradlew build}
\usage{
oml_build(oml_repository)
}
\arguments{
\item{oml_repo}{a repository of the oml project}
}
\value{
output from gradle tasks
}
\description{
R wrapper for ./gradlew build
}
