% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oml_clean.R
\name{oml_clean}
\alias{oml_clean}
\title{R wrapper for ./gradlew clean}
\usage{
oml_clean(oml_repository)
}
\arguments{
\item{oml_repo}{a repository of the oml project}
}
\value{
output from gradle tasks
}
\description{
R wrapper for ./gradlew clean
}
