% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oml_owlLoad.R
\name{oml_owlLoad}
\alias{oml_owlLoad}
\title{R wrapper for ./gradlew owlLoad}
\usage{
oml_owlLoad(oml_repository)
}
\arguments{
\item{oml_repository}{}
}
\value{
output from gradle tasks
}
\description{
R wrapper for ./gradlew owlLoad
}
