% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oml_owlQuery.R
\name{oml_owlQuery}
\alias{oml_owlQuery}
\title{R wrapper for ./gradlew owlQuery}
\usage{
oml_owlQuery(oml_repository)
}
\arguments{
\item{oml_repository}{}
}
\value{
output from gradle tasks
}
\description{
R wrapper for ./gradlew owlQuery
}
