% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oml_owlReason.R
\name{oml_owlReason}
\alias{oml_owlReason}
\title{R wrapper for ./gradlew owlReason}
\usage{
oml_owlReason(oml_repository)
}
\arguments{
\item{oml_repository}{}
}
\value{
output from gradle tasks
}
\description{
R wrapper for ./gradlew owlReason
}
