% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oml_refresh.R
\name{oml_refresh}
\alias{oml_refresh}
\title{R wrapper for refresh fuseki servers.}
\usage{
oml_refresh(oml_repository)
}
\arguments{
\item{oml_repository}{}
}
\value{
output from gradle tasks
}
\description{
R wrapper for refresh fuseki servers.
}
