% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oml_startFuseki.R
\name{oml_startFuseki}
\alias{oml_startFuseki}
\title{R wrapper for ./gradlew startFuseki}
\usage{
oml_startFuseki(oml_repository)
}
\arguments{
\item{oml_repository}{}
}
\value{
output from gradle tasks
}
\description{
R wrapper for ./gradlew startFuseki
}
