% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oml_stop_Daemon.R
\name{oml_stop_Daemon}
\alias{oml_stop_Daemon}
\title{R wrapper for ./gradlew --stop}
\usage{
oml_stop_Daemon(oml_repository)
}
\arguments{
\item{oml_repository}{}
}
\value{
output from gradle tasks
}
\description{
R wrapper for ./gradlew --stop
}
