% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oml_stopFuseki.R
\name{oml_stopFuseki}
\alias{oml_stopFuseki}
\title{R wrapper for ./gradlew stopFuseki}
\usage{
oml_stopFuseki(oml_repository)
}
\arguments{
\item{oml_repository}{}
}
\value{
output from gradle tasks
}
\description{
R wrapper for ./gradlew stopFuseki
}
